bookitoBackend/requirements.txt
Django
djangorestframework
knox
djnago-rest-knox
